refactor(LottoManager) : 복잡한 로또 계층 통합
