--- src/main/java/lotto/domain/LottoMachine.java
@@ -1,24 +1,0 @@
-package lotto.domain;
-
-import lotto.domain.lotto.Lotto;
-import lotto.domain.lotto.Lottos;
-import lotto.util.LottoNumberGenerator;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class LottoMachine {
-    private final LottoNumberGenerator generator;
-
-    public LottoMachine(LottoNumberGenerator generator) {
-        this.generator = generator;
-    }
-
-    public Lottos createLottos(int count) {
-        List<Lotto> lottos = new ArrayList<>();
-        for (int i = 0; i < count; i++) {
-            lottos.add(new Lotto(generator.generate()));
-        }
-        return new Lottos(lottos);
-    }
-}
